Change fully_shard memory profiler dtype from float32 to bfloat16
Match the recipe training scripts which use bfloat16 for model params.

+++ bionemo-recipes/recipes/llama3_native_te/benchmarks/fully_shard_memory_profile.py
@@ -59,7 +59,7 @@ NUM_LAYERS = 3
 SEQ_LEN = 32
 BATCH_PER_RANK = 2
 NUM_STEPS = 5
-DTYPE = torch.float32
+DTYPE = torch.bfloat16
 
 
 def dist_print(msg):
